+++ packages/webview_flutter/webview_flutter_platform_interface/lib/src/method_channel/webview_method_channel.dart
@@ -216,6 +216,12 @@ class MethodChannelWebViewPlatform implements WebViewPlatformController {
         .then<bool>((dynamic result) => result! as bool);
   }
 
+  /// Method channel implementation for [WebViewPlatform.setCookie].
+  static Future<void> setCookie(WebViewCookie cookie) {
+    return _cookieManagerChannel.invokeMethod<void>(
+        'setCookie', cookie.toJson());
+  }
+
   static Map<String, dynamic> _webSettingsToMap(WebSettings? settings) {
     final Map<String, dynamic> map = <String, dynamic>{};
     void _addIfNonNull(String key, dynamic value) {
@@ -261,6 +267,9 @@ class MethodChannelWebViewPlatform implements WebViewPlatformController {
       'autoMediaPlaybackPolicy': creationParams.autoMediaPlaybackPolicy.index,
       'usesHybridComposition': usesHybridComposition,
       'backgroundColor': creationParams.backgroundColor?.value,
+      'cookies': creationParams.cookies
+          .map((WebViewCookie cookie) => cookie.toJson())
+          .toList()
     };
   }
 }